fix(release): move repo-wide-suites out of the dockerignored test/ directory
The web-platform image build has failed on every merge to main since
2026-08-20, so the deploy job is skipped and production keeps serving
dc201e757f63. Measured this session: /health reports build_sha
dc201e757f63faa2001b4cf3e4ae4d8e6748bb38.

The fatal error, from the run-32860030881 build log:

    Failed to compile.
    Type error: Cannot find module './test/repo-wide-suites' or its
    corresponding type declarations.
    > 4 | import { REPO_WIDE_SUITES } from "./test/repo-wide-suites";

vitest.config.ts is present in the Docker build context and is
type-checked by next build; test/ is excluded by .dockerignore line 106.
The import crosses that boundary, so the type-check cannot resolve it.

Reproduced independently of CI with a busybox context probe:
vitest.config.ts PRESENT, test/repo-wide-suites.ts ABSENT.

The fix moves the file to the app root rather than re-including test/ in
the build context. The move deletes the cross-boundary edge; a
re-include would create a permanent exception that then needs guarding
forever. It also matches how the structurally identical earlier case was
resolved: the edge was removed, not re-included.

The file does not reach the runtime image. The runner stage copies no
app-root file except next.config.mjs, so the move adds nothing to the
shipped artifact.

No test-runner blast radius: nothing at app root is globbed by any
vitest project. unit is test/** plus lib/**, component is
test/**/*.test.tsx, and repo-wide is the explicit manifest.

Reference set updated at all five sites, enumerated by git grep:
vitest.config.ts:4, repo-wide-containment.test.ts:42 plus its two
operator-facing messages at :210 and :232, and the comment in
scripts/lib/test-relevance-paths.sh:248. The prose fixes matter because
those strings are what the guard prints to whoever trips it.

Verified: busybox probe now reports PRESENT, and tsc --noEmit is clean.

Ref #7395.
